Put SlashJitpack stuff in services
